requirements.txt
langgraph
google-genai
pydantic